requirements.txt
flask>=2.0
pyserial
pytest
